requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
mistralai>=1.0.0
pydantic>=2.5.2
python-dateutil==2.8.2
icalendar==5.0.11
jinja2==3.1.2
aiofiles==23.2.1
